add description of change to rolled
rolled used to return only the array for the "main roll", ignoring any rolls which occurred as a result of calls by the parser to process rolls that are added to the main roll.  Now compound rolls will be reflected as an array of arrays.  Single rolls still return a regular array, instead of returning an array that contains only a single element that is also an array